3rd try

+++ src/main/java/jdbc/SimpleJDBCRepository.java
@@ -28,18 +28,18 @@ public class SimpleJDBCRepository {
     private static final String findAllUserSQL = "select * from myusers";
 
     public Long createUser() {
-        int res;
+        Long res;
         connection = new CustomConnector().getConnection(CustomDataSource.getInstance().getUrl(), CustomDataSource.getInstance().getName(), CustomDataSource.getInstance().getPassword());
         try {
             ps = connection.prepareStatement(createUserSQL);
             ps.setString(1, "Jasur");
             ps.setString(2, "Rahmonov");
-            ps.setInt(3, 27);
-            res = ps.executeUpdate();
+            ps.setLong(3, 27L);
+            res = (long) ps.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
-        return (long) res;
+        return res;
     }
 
     public User findUserById(Long userId) {
